separated Hero into About section
