+++ src/App.jsx
@@ -1,12 +1,13 @@
 import { BrowserRouter } from "react-router-dom";
 
-import { Hero, Navbar } from "./components";
+import { Hero, Navbar, About } from "./components";
 
 function App() {
   return (
     <BrowserRouter>
       <div className="relative z-0 bg-blackbean">
         <Hero />
+        <About />
       </div>
     </BrowserRouter>
   );
