+++ src/components/index.js
@@ -1,4 +1,5 @@
 import Hero from "./Hero";
 import Navbar from "./Navbar";
+import About from "./About";
 
-export { Hero, Navbar };
+export { Hero, Navbar, About };
